backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.0
pydantic-settings==2.5.2
celery==5.4.0
redis==5.1.1
spacy==3.7.6
scikit-learn==1.5.2
sentence-transformers==3.1.1
pdfminer.six==20231228
python-docx==1.1.2
numpy==1.26.4
pandas==2.2.3
httpx==0.27.2
python-dotenv==1.0.1